requirements.txt
asgiref==3.7.2
Brotli==1.1.0
dj-database-url==2.1.0
Django==4.2.5
djangorestframework==3.14.0
gunicorn==21.2.0
install==1.3.5
packaging==23.1
psycopg2-binary==2.9.7
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.5.0
